+++ src/main/java/com/barlinc/unusual_prehistory/entity/mob/future/Cotylorhynchus.java
@@ -101,7 +101,7 @@ public class Cotylorhynchus extends PrehistoricMob {
     @Nullable
     @Override
     public AgeableMob getBreedOffspring(@NotNull ServerLevel level, @NotNull AgeableMob mob) {
-        return UP2Entities.COTYLORHYNCHUS.get().create(level);
+        return UP2Entities.LYSTROSAURUS.get().create(level);
     }
 
     @Nullable